Begin Performance Charting JavaScript
This commit pulls in RequireJS and several plugins for it (goog, async, propertyParser) as dependencies.

JavaScript is currently very basic: Draws a simple chart with data obtained via AJAX.  Doesn't even translate the dates from UNIX timestamps to actually dates before display (this makes the horizontal axis of the chart difficult to decipher).