Performance Charting JavaScript - Dates
The JavaScript front-end for performance charting now outputs a Google chart with dates along the horizontal axis.

This pulls in Moment.js and Moment Timezone as dependencies.

Also: The back-end now sends the survey's timezone in its response as the "timezone" key of the JSON object which forms the response's body.