Performance Charting Improvements
The performance charting chart now:

-	Fills its parent container (doesn't resize as the page does however)
-	Always has an appropriate scale on its vertical axis (i.e. it is clamped to the possible range of the data rather than the actual range)

The front-end code in app.js now waits for page load to run (just in case).

I attempted to get the chart to load/render on page load but this causes an error inside the Google Chart API: "Cannot read property 'length' of null" which seems to be linked to the fact that the chart isn't visible on page load.

app.js no longer calls console.log.