Performance Charting Front-End - Draw Chart on Tab Show
Having the performance charting JavaScript component always running was problematic due to the fact Google Charts doesn't seem to like drawing charts in elements which are hidden (as the "Performance Charting" tab is on page load).

To get around this the code in app.js now enables the performance charting manager when its tab is shown and disables it just before its tab is hidden.

In implementing this I discovered an interesting bug/problem due to the combination of "legacy" JavaScript code and AMD-aware code:

1.	jQuery was loaded through its hardcoded script tag in the header
2.	Bootstrap was loaded through its hardcoded script tag in the header
3.	Bootstrap attached itself to the jQuery from #1
4.	RequireJS was loaded and began resolving dependencies for/running AMD-aware code
5.	The code in app.js requires jQuery, therefore RequireJS loaded a second copy of jQuery
6.	When the code in app.js executed it used the jQuery from #5 which didn't have Bootstrap attached

As a result of the above the "shown.bs.tab" and "hide.bs.tab" events never seemed to fire.

The solution was to create a dummy jquery.js which is AMD-aware and simply returns a reference to the global jQuery (i.e. the jQuery loaded in #1 above).  This eliminates the second copy of jQuery from #5 above as RequireJS loads jquery.js (which provides the global jQuery from #1 above) rather than a whole new copy of jQuery.