\Fgms\Bundle\SurveyBundle\Controller\DefaultController - Summary
The ::chartAction and ::chartRangeAction methods now also responds with the "summary" key which contains a summary of the selected data set.

For "rating" type questions this is a JSON object with the following properties:

-	"perfect": The number of entries with a rating of 5
-	"acceptable": The number of entries with a rating less than 5 but greater than or equal to the threshold/trigger for the question
-	"unacceptable": The number of entries with a rating less than the threshold/trigger for the question
-	"total": The total number of entries

For "polar" type questions this is a JSON object with the following properties:

-	"good": The number of positive answers
-	"bad": The number of negative answers
-	"total": The total number of answers

For "open" type questions this is a JSON object with the following properties:

-	"good": The number of entries with answers
-	"bad": The number of entries without answers
-	"total": The total number of entries

This commit also fixes a bug which was causing \Fgms\Bundle\SurveyBundle\Entity\Questionnaire objects with a NULL value for an "open" type question to be discounted as invalid: Apparently this is how unanswered entries are represented.