Guest Feedback Front-End - Improve Table Display
The first and third columns now have a fixed width which prevents them from being squashed by the lengthy content in the middle column.

Also: Even though "Posted to Website?" data is not yet available/implemented a <td> is now still generated which causes the table borders etc. to be consistent/full width.