Performance Charting Back-End - Add "negative" to JSON Object
The JSON object returned by the performance charting back-end now includes a "negative" property if the question being queried is of "polar" type.  This indicates whether the sense of the "polar" question is inverted.  I.e. if this value is false then "yes" is good and "no" is bad otherwise vice versa.