\Fgms\Bundle\SurveyBundle\Utility\DoctrineRand & \Fgms\Bundle\SurveyBundle\Controller\DefaultController::testimonialsAction Random Support
The \Fgms\Bundle\SurveyBundle\Utility\DoctrineRand class adds a RAND() function to Doctrine DQL.  It must be enabled by placing the following in the Symfony config.yml:

doctrine:
    orm:
        dql:
            numeric_functions:
                Rand: Fgms\Bundle\SurveyBundle\Utility\DoctrineRand

There may be a better way to enable it (using something local to this bundle rather than going out to the global Symfony config.yml) but I do not know of it at this time.

With this class/DQL function in place support for selecting random testimonials has been added to \Fgms\Bundle\SurveyBundle\Controller\DefaultController::testimonialsAction: Pass "random" as the $order parameter.